8-print_array.c

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -8,8 +8,8 @@
  */
 int _atoi(char *s)
 {
-	int sign = 1;
-	int result = 0;
+	long int sign = 1;
+	long int result = 0;
 
 	while (*s)
 	{
